+++ packages/infrastructure/modules/vpc/data.tf
@@ -1,0 +1,1 @@
+data "aws_region" "current" {}
