+++ packages/infrastructure/modules/vpc/main.security_group.tf
@@ -1,0 +1,10 @@
+resource "aws_security_group" "appsync_endpoint" {
+  vpc_id = aws_vpc.main.id
+  
+  ingress {
+    from_port   = 443
+    to_port     = 443
+    protocol    = "tcp"
+    cidr_blocks = [aws_vpc.main.cidr_block]
+  }
+}
