+++ packages/infrastructure/modules/vpc/main.tf
@@ -1,0 +1,5 @@
+resource "aws_vpc" "main" {
+  cidr_block = "10.0.0.0/16"
+  enable_dns_support = true
+  enable_dns_hostnames = true
+}
